requirements.txt
Django
djangorestframework
mysqlclient  
